+++ mkchromecast/__init__.py
@@ -61,7 +61,13 @@ Example:
 
 ''')
 parser.add_argument('-n', '--name', action="store_true", help='Use this option if you know the name of the Google Cast you want to connect')
-parser.add_argument('-r', '--reset', action="store_true", help='When the application fails, and you have no audio in your computer, use this option to reset')
+parser.add_argument('-r', '--reset', action="store_true", help='''
+When the application fails, and you have no audio in your computer, use this
+option to reset the computer's audio
+''')
+parser.add_argument('--reboot', action="store_true", help='''
+Reboot the Google Cast device
+''')
 parser.add_argument('-s', '--select-cc', action="store_true", help='If you have more than one Google Cast device use this option')
 parser.add_argument('--sample-rate', type=int, default='44100', help=
 '''
@@ -149,6 +155,23 @@ if args.reset == True:
         remove_sink()
     terminate()
 
+if args.config == True or args.discover == True or args.name == True:
+    print (colors.error('This option is not implemented yet.'))
+    sys.exit(0)
+
+"""
+Reboot
+"""
+if args.reboot == True:
+    print (colors.error('This option is not implemented yet.'))
+    sys.exit(0)
+else:
+    print (colors.error('This option is not implemented yet.'))
+    sys.exit(0)
+
+"""
+Not yet implemented
+"""
 if args.config == True or args.discover == True or args.name == True:
     print (colors.error('This option is not implemented yet.'))
     sys.exit(0)