+++ changelog.md
@@ -3,6 +3,10 @@
     - This release fixes the system tray menu for Linux.
     - Now avconv is supported.
     - Pass the `--update` option to update the repository.
+    - It is now possible to control the Google cast device volume from the
+      systray.
+    - You can reboot the Google cast devices from the systray and the command
+      line.
 
 * mkchromecast (0.2.3.1) **released**: 2016/05/08